% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{prepare_reference_ensembl}
\alias{prepare_reference_ensembl}
\title{Prepare reference genome data directory using Ensembl data}
\usage{
prepare_reference_ensembl(output_dir, fasta, gtf)
}
\arguments{
\item{output_dir}{Output directory path}

\item{fasta}{A path or an URL to an Ensembl FASTA file contaning genome sequence}

\item{gtf}{A path or an URL to an Ensembl GTF file contaning genome annotations}
}
\description{
Prepares a directory containing reference genome files based on Ensembl FASTA and GTF files.
The output directory contains following files:
\itemize{
\item genome.fasta - genome sequence
\item annotations.gtf - genome annotations
\item transcriptome.fasta - transcriptome sequence
\item genes.rds - gene GRanges
\item genes.csv - gene descriptions
\item transcripts.rds - transcript GRanges
\item transcripts.csv - transcript descriptions
}
}
