% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{prepare_reference_kallisto_rna_velocity}
\alias{prepare_reference_kallisto_rna_velocity}
\title{Add files for kallisto RNA velocity analysis to the reference genome data directory}
\usage{
prepare_reference_kallisto_rna_velocity(output_dir)
}
\arguments{
\item{output_dir}{Reference genome data directory}
}
\description{
Adds files for kallisto RNA velocity analysis to the reference genome data directory.
Creates a kallisto_rna_velocity/ subdirectory containing following files:
\itemize{
\item cDNA_introns.fa - spliced transcripts and flanking intronic sequences
\item cDNA_tx_to_capture.txt - IDs of spliced transcripts
\item introns_tx_to_capture.txt - IDs of intronic sequences
\item tr2g.txt - a table matching spliced transcript and intronic sequence IDs to gene IDs
}
}
