% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdb.R
\name{txdb_to_gtf}
\alias{txdb_to_gtf}
\title{Export a TxDb object to a clean GTF file}
\usage{
txdb_to_gtf(txdb, file)
}
\arguments{
\item{txdb}{A TxDb object}

\item{file}{Output file path}
}
\description{
Exports a TxDb object to a clean GTF file (i.e. only 'exon' and 'CDS' features
and 'gene_id' and 'transcript_id' tags)
}
